refactor: move fiterDependentParameters to advanced settings general utils
