+++ application/ui/src/features/models/model-listing/model-training-parameters/model-training-parameters.component.tsx
@@ -7,7 +7,7 @@ import { Grid, Text } from '@geti/ui';
 
 import { TrainingConfigurationParameter } from '../../../../constants/shared-types';
 import { useGetModelTrainingConfiguration } from '../../hooks/api/use-get-model-training-configuration.hook';
-import { filterDependentParameters } from '../../train-model/advanced-settings/training/learning-parameters/utils';
+import { filterDependentParameters } from '../../train-model/advanced-settings/utils';
 import { Box } from '../components/box/box.component';
 import { findGroupByKey, flattenParameters } from './utils';
