+++ application/ui/src/features/models/train-model/advanced-settings/utils.test.ts
@@ -11,9 +11,10 @@ import {
     ConfigurableParameterGroup,
     TrainingConfigurationParameter,
 } from '../../../../constants/shared-types';
-import { isParameterGroup } from '../../model-listing/model-training-parameters/utils';
+import { isParameter, isParameterGroup } from '../../model-listing/model-training-parameters/utils';
 import { learningParameters } from './training/learning-parameters/mocks';
-import { deepReplaceParameters } from './utils';
+import { LearningConfigurationGroup } from './training/learning-parameters/utils';
+import { deepReplaceParameters, filterDependentParameters } from './utils';
 
 const getParam = (params: ConfigurableParameter[], key: string): ConfigurableParameter | undefined =>
     params.find((p) => p.key === key) as ConfigurableParameter | undefined;
@@ -307,5 +308,76 @@ describe('deepReplaceParameters', () => {
             const replaced = getParam(result as ConfigurableParameter[], 'weight_decay');
             expect(replaced?.value).toBe(0.001);
         });
+    });
+});
+
+describe('filterDependentParameters', () => {
+    it('keeps independent parameters in their original order', () => {
+        const result = filterDependentParameters(learningParameters.parameters);
+        const keys = result.filter(isParameter).map((p) => p.key);
+
+        expect(keys).toContain('max_epochs');
+        expect(keys).toContain('batch_size');
+        expect(keys.indexOf('max_epochs')).toBeLessThan(keys.indexOf('batch_size'));
+    });
+
+    it('places dependent parameters immediately after the parameter they depend on', () => {
+        const schedulerGroup = learningParameters.parameters.find(
+            (p) => isParameterGroup(p) && p.key === 'scheduler'
+        ) as LearningConfigurationGroup;
+        const result = filterDependentParameters(schedulerGroup.parameters);
+        const keys = result.filter(isParameter).map((p) => p.key);
+
+        const typeIndex = keys.indexOf('type');
+        const factorIndex = keys.indexOf('factor');
+        const patienceIndex = keys.indexOf('patience');
+
+        // 'factor' and 'patience' depend on 'type' === 'reduce_lr_on_plateau' (which is the current value)
+        // so they should immediately follow 'type'
+        expect(factorIndex).toBe(typeIndex + 1);
+        expect(patienceIndex).toBe(typeIndex + 2);
+    });
+
+    it('excludes dependent parameters whose condition does not match the current value', () => {
+        const schedulerGroup = learningParameters.parameters.find(
+            (p) => isParameterGroup(p) && p.key === 'scheduler'
+        ) as LearningConfigurationGroup;
+        const result = filterDependentParameters(schedulerGroup.parameters);
+        const keys = result.filter(isParameter).map((p) => p.key);
+
+        // 'min_lr' depends on type === 'cosine_annealing', but type === 'reduce_lr_on_plateau'
+        expect(keys).not.toContain('min_lr');
+    });
+
+    it('preserves parameter groups in the output', () => {
+        const result = filterDependentParameters(learningParameters.parameters);
+        const groupKeys = result.filter(isParameterGroup).map((p) => p.key);
+
+        expect(groupKeys).toContain('early_stopping');
+        expect(groupKeys).toContain('scheduler');
+        expect(groupKeys).toContain('gradient_accumulation');
+        expect(groupKeys).toContain('gradient_clip');
+    });
+
+    it('returns an empty array when given an empty array', () => {
+        expect(filterDependentParameters([])).toEqual([]);
+    });
+
+    it('returns a single independent parameter unchanged', () => {
+        const param = getMockedConfigurationParameter({ value_type: 'int', key: 'solo', depends_on: null });
+        const result = filterDependentParameters([param]);
+
+        expect(result).toEqual([param]);
+    });
+
+    it('does not include standalone dependent parameters that have no matching parent', () => {
+        const dependent = getMockedConfigurationParameter({
+            value_type: 'float',
+            key: 'orphan',
+            depends_on: { some_key: 'some_value' },
+        });
+        const result = filterDependentParameters([dependent]);
+
+        expect(result).toEqual([]);
     });
 });
